Config

Config files are YAML files with the information required to make Oanda API Calls A sample config file is as follows:

hostname: <<api-fxpractice.oanda.com or api-fxtrade.oanda.com>>
streaming_hostname: <<stream-fxpractice.oanda.com or stream-fxtrade.oanda.com>>
token: <<your token>>
username: <<Account username>>
account: <<Account to initially interact with>>

Functions

Julianda.Config.loadConfig — Function

loadConfig(path::String)

Loads a YAML config

Arguments

path: The relative path to the config file

Julianda.Config.saveConfig — Function

saveConfig(path::String, config::config)

Saves a config as a YAML file

Arguments

path: Path to save the config at

config: The config struct to save

Julianda.Config.changeAccount — Function

changeAccount(config::config, acc::string)

Returns the config with the new account number

Arguments

conf: The config file to update

acc: The string of the account identifier