Instruments

Functions

Julianda.Instrument.getCandles — Function

getCandles(config::config, instrument::String, lastn::Int = 10, price::String = "M", granularity::String = "M5";kwargs...)
getCandles(config::config, instrument::String, from::DateTime, to::DateTime, price::String = "M", granularity::AbstractString = "M5";kwargs...)
getCandles(config::config, instrument::String, from::DateTime, n::Int = 10, price::String = "M", granularity::AbstractString = "M5";kwargs...)
getCandles(config::config, instrument::String, n::Int, to::DateTime, price::String = "M", granularity::AbstractString = "M5";kwargs...)
getCandles(config::config, instrument::String, from::DateTime, price::String = "M", granularity::AbstractString = "M5";kwargs...)

Get candle information of a given instrument and returns a Candle struct Information includes: time, granularity, open, high, low, close, volume and a complete indicator

getCandles has five methods depending on how the candles to retrieve are selected

lastn: last "n" candles

from and to: candles in a time interval specified by two dates

from and "n", to and "n": n candles from or to the specified date

from: all candles from the specified date

Arguments

'config::config': a valid struct with user configuracion data

'instrument::String": a valid instrument (e.g. "EUR_USD")

'price::String': "A" for ask, "B" for bid, "M" for medium or a combination ot them

'granularity::String': a valid time interval ["S5","S10","S15","S30","M1","M2","M4","M5","M10","M15","M30","H1","H2","H3","H4","H6","H8","H12","D","W","M"]

Keyword Arguments

'smooth::Bool'

'includeFirst::Bool'

'dailyaligment::Int'

'alignmentTimezone::String'

'weeklyAlignment::String'

Examples

getCandles(userdata,"EUR_USD",10,"A","M30")
getCandles(userdata,"EUR_JPY",DateTime(2019,1,1),DateTime(2019,1,31),"B","H1")
getCandles(userdata,"EUR_USD",DateTime(2019,1,31),10,"A","M30")
getCandles(userdata,"EUR_CHF",10,DateTime(2019,1,31),"AB","M5")
getCandles(userdata,"EUR_USD",DateTime(2019,1,31),"M","D")

Julianda.Instrument.getOrderBook — Function

getOrderBook(config::config,instrument::String,time::DateTime=now())

Example

getOrderBook(userdata,"EUR_CHF",DateTime(2017,1,31,4,00))

Julianda.Instrument.getPositionBook — Function

getPositionBook(config::config,instrument::String,time::DateTime=now())

Example

getPositionBook(userdata,"EUR_CHF",DateTime(2017,1,31,4,00))