Orders

Functions

Julianda.Order.marketOrder — Function

marketOrder(config, instrument, units;[TIF, positionFill, priceBound, TP ,SL ,tSL, clientExt ,tradeExt])

Places a market order

#Arguments

'config::config': a valid struct with user configuration data

'instrument::String': a valid instrument (e.g. "EUR_USD")

'units::String': units to buy (positive value) or sell (negative value)

Keyword Arguments

'TIF::String' : timeinForce value ("GTC","GTD","GFD","FOK","IOC")

'positionFill::String' : how positions are modified when the order is filled ("OPENONLY", "REDUCEFIRST", "REDUCE_ONLY","DEFAULT"). Defaults to "DEFAULT"

'priceBound::String': worst price for filling

'TP', 'SL', 'tSL' : a NamedTuple with order-on-fill details. Valid values are 'price::Real', 'distance::Real', 'timeInForce::String' and 'gtdTime::String'

'clientExt'(not implemented)

'tradeExt' (not implemented)

#Examples

marketOrder(userData,"EURJPY",100) marketOrder(userData,"EURCHF",100,SL=(distance=0.1,),TP=(price=1.12,),tSL=(distance=0.3,)) #Do not forget the comma for single value NamedTuple

Missing docstring.

Missing docstring for Order.nonMarketOrder. Check Documenter's build log for details.

Julianda.Order.limitOrder — Function

limitOrder(config, instrument, units, price;[TIF, positionFill, priceBound, TP ,SL ,tSL, clientExt ,tradeExt])

Places a limit Order

#Arguments

'config::config': a valid struct with user configuration data

'instrument::String': a valid instrument (e.g. "EUR_USD")

'units::String': units to buy (positive value) or sell (negative value)

'price::Real'

Keyword Arguments

'TIF::String' : timeinForce value ("GTC","GTD","GFD","FOK","IOC")

'positionFill::String' : how positions are modified when the order is filled ("OPENONLY", "REDUCEFIRST", "REDUCE_ONLY","DEFAULT"). Defaults to "DEFAULT"

'priceBound::String': worst price for filling

'TP', 'SL', 'tSL' : a NamedTuple with order-on-fill details. Valid values are 'price::Real', 'distance::Real', 'timeInForce::String' and 'gtdTime::String'

'clientExt'(not implemented)

'tradeExt' (not implemented)

#Examples

limitOrder(userData,"EURUSD",100, 1.10) limitOrder(userData,"EURJPY",100,117,SL=(distance=1,),TP=(price=12,),tSL=(distance=3,)) #Do not forget the comma for single value NamedTuple

Julianda.Order.stopOrder — Function

stopOrder(config, instrument, units, price;[TIF, positionFill, priceBound, TP ,SL ,tSL, clientExt ,tradeExt])

Places a stop Order

#Arguments

'config::config': a valid struct with user configuration data

'instrument::String': a valid instrument (e.g. "EUR_USD")

'units::String': units to buy (positive value) or sell (negative value)

'price::Real'

Keyword Arguments

'TIF::String' : timeinForce value ("GTC","GTD","GFD","FOK","IOC")

'positionFill::String' : how positions are modified when the order is filled ("OPENONLY", "REDUCEFIRST", "REDUCE_ONLY","DEFAULT"). Defaults to "DEFAULT"

'priceBound::String': worst price for filling

'TP', 'SL', 'tSL' : a NamedTuple with order-on-fill details. Valid values are 'price::Real', 'distance::Real', 'timeInForce::String' and 'gtdTime::String'

'clientExt'(not implemented)

'tradeExt' (not implemented)

#Examples

stopOrder(userData,"EURUSD",100, 1.10) stopOrder(userData,"EURJPY",100,117,SL=(distance=1,),TP=(price=12,),tSL=(distance=3,)) #Do not forget the comma for single value NamedTuple

Julianda.Order.marketIfTouchedOrder — Function

marketIfTouchedOrder(config, instrument, units, price;[TIF, positionFill, priceBound, TP ,SL ,tSL, clientExt ,tradeExt])

Places a market-if-touched Order

#Arguments

'config::config': a valid struct with user configuration data

'instrument::String': a valid instrument (e.g. "EUR_USD")

'units::String': units to buy (positive value) or sell (negative value)

'price::Real'

Keyword Arguments

'TIF::String' : timeinForce value ("GTC","GTD","GFD","FOK","IOC")

'positionFill::String' : how positions are modified when the order is filled ("OPENONLY", "REDUCEFIRST", "REDUCE_ONLY","DEFAULT"). Defaults to "DEFAULT"

'priceBound::String': worst price for filling

'TP', 'SL', 'tSL' : a NamedTuple with order-on-fill details. Valid values are 'price::Real', 'distance::Real', 'timeInForce::String' and 'gtdTime::String'

'clientExt'(not implemented)

'tradeExt' (not implemented)

#Examples

marketifTouchedOrder(userData,"EURuSD",100, 1.10) marketifTouchedOrder(userData,"EURJPY",100,117,SL=(distance=1,),TP=(price=12,),tSL=(distance=3,)) #Do not forget the comma for single value NamedTuple

Julianda.Order.relatedOrder — Function

 relatedOrder(config, tradeID, type, detail)

Places a stop loss, take profit or trailing stop loss to an open trade (not order)

#Arguments

'config::config': a valid struct with user configuration data

'tradeID::Union{Int,String}': a valid open trade ID (e.g. "123")

'type::String' : type of the order "STOPLOSS", "TAKEPROFIT" or "TRAILINGSTOPLOSS"

'details::NamedTuple': a NamedTuple with order-on-fill details. Valid values are 'price::Real', 'distance::Real', 'timeInForce::String' and 'gtdTime::String'

#Examples

relatedOrder(config, "156", "STOP_LOSS", details=(distance=0.1,)) #Do not forget the comma for single value NamedTuple

Julianda.Order.getOrders — Function

getOrders(config, count::Int=50; kwargs...) getOrders(config, IDlist::Vector; kwargs...)

#Arguments

'config::config': a valid struct with user configuration data

'count::Int': number of orders to return

'IDlist::Vector{String}' : a vector of valid order IDs

#Keyword Arguments

'state::String: a string with values "PENDING", "FILLED", "TRIGGERED", "CANCELLED" or "ALL". Defaults to "PENDING"

'instrument::String'

'beforeID::String': last ID to retrieve

#Examples getOrders(userData,5) getOrders(userData,25, state="FILLED", instrument="EUR_USD") getOrders(userData,["123","234","345"])

Julianda.Order.getPendingOrders — Function

getPendingOrders(config)

#Examples getPendingOrders(userData)

Julianda.Order.getOrder — Function

getOrder(config, orderID)

#Arguments

'config::config': a valid struct with user configuration data

'orderID::Union{String,Int}' : avalid order IDs

#Examples getOrder(userdata,"100")

Julianda.Order.replaceOrder — Function

replaceOrder(config, ID, instrument, units;[TIF, positionFill, priceBound, TP ,SL ,tSL, clientExt ,tradeExt])

#Examples

replaceOrder(userData,165,"LIMIT","EURJPY",100) replaceOrder(userData,170,"EURCHF",100,SL=(distance=0.1,),TP=(price=1.12,),tSL=(distance=0.3,))

Julianda.Order.cancelOrder — Function

cancelOrder(config, ID::Union{String,Int})

#Arguments

'config::config': a valid struct with user configuration data

'orderID::Union{String,Int}' : avalid order IDs

#Examples cancelOrder(userdata,"100")

Julianda.Order.orderClientExtensions — Function

orderClientExtensions(config::config, tradeID::String; clientID::String="", tag::String="", comment::String="")

Lets add user information to a specific Trade

Arguments

'config::config': a valid struct with user configuracion data

'tradeID::string': a valid trade ID

clientID, tag and comment: strings with the user information

Example

orderClientExtensions(userconfig,"66", clientID="007",tag="foo")

Structs

Missing docstring.

Missing docstring for Order.extensions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order.takeProfit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order.stopLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order.trailingStopLoss. Check Documenter's build log for details.

Julianda.Order.orderRequest — Type

Detailed OrderRequest struct from Oanda