Pricing

Functions

OandaTrade.Pricing.getPrice — Function

function getPrice(config, instruments)

Get the most recent price update of an instrument

Returns an object of type 'priceData'

Arguments

- 'config::config': a valid struct with user configuracion data
- 'instruments::Vector{String}': a vector of valid pairs (e.g. ["EUR_USD","EUR_JPY"])

#Example

getPrice(userconfig, ["EUR_USD","EUR_JPY"])

OandaTrade.Pricing.checkMarket — Function

Get the most recent price update of an instrument

OandaTrade.Pricing.streamPrice — Function

function streamprice(f, config , instruments)

Returns a stream of price objects and apply a function to each one of them

Arguments

- 'f::Function': a function to apply to each price struct object. Streamprice accepts do block format
- 'config::config': a valid struct with user configuracion data
- 'instruments::Vector{String}': a vector of valid pairs (e.g. ["EUR_USD","EUR_JPY"])

#Example

streamprice(userconfig, ["EUR_JPY"]) do price
    println(price)
end

Missing docstring.

Missing docstring for Pricing.getLatestCandles. Check Documenter's build log for details.

Structs

OandaTrade.Pricing.priceBucket — Type

Ask / Bid pricing data

Arguments

'price': Price of the ask / bid

'liquidity': Liquidity of the ask / bid

OandaTrade.Pricing.price — Type

Pricing data of an instrument

Arguments

'type': Type

'instrument': Instrument of the price

'time': Time of the price update

'bids::Vector{priceBucket}': bid info

'asks::Vector{priceBucket}': ask info

'closeoutBid': Closeout bid price

'closeoutAsk': Closeout ask price

'tradeable': Can you trade this instrument

OandaTrade.Pricing.candlestickdata — Type

Candlestick tick data

Fields

'o': Opening price

'h': Highest price

'l': Lowest price

'c': Closing price

OandaTrade.Pricing.candlestick — Type

Candlesticks

Fields

'time': The time the data corresponds to

'bid::candlestickdata': Candlestick tick data for bids

'ask::candlestickdata': Candlestick tick data for asks

'mid::candlestickdata': Candlestick tick data for the mid

'volume': Volume of trades

'complete': is this complete data?

OandaTrade.Pricing.candles — Type

Candles

Fields

'instrument': Instrument the data relates to

'granularity': The granularity of the candle data

'candles::Vector{candlestick}': The candlestick data