Trades

Functions

Julianda.Trade.getTrades — Function

getTrades(config::config, instrument::String, state::String="OPEN", count::Int=50; kwargs...)

Return an array of trade struct

Arguments

'config::config': a valid struct with user configuracion data

'state::String": the state to filter the requested Trades by (OPEN, CLOSED, CLOSEWHENTRADEABLE, ALL)

'count::Int': max number of trades to return

Keyword Arguments

'instrument::String": a valid instrument (e.g. "EUR_USD")

'ids::String' List of trades to retrieve as ID values separated by commas

'beforeID::String' The maximum trade ID to return

Examples

getTrades(userConfig,"CLOSED";instrument="EUR_USD")

Julianda.Trade.getOpenTrades — Function

getOpenTrades(config::config)

Return an array of trade struct

Arguments

'config::config': a valid struct with user configuracion data

Examples

getOpenTrades(userconfig)

Julianda.Trade.getTrade — Function

getTrade(config::config, tradeID::String)

Return a specific trade

Arguments

'config::config': a valid struct with user configuracion data

'id::string': a valid trade ID

Examples

getTrades(userconfig,"66")

Julianda.Trade.closeTrade — Function

closeTrade(config::config, tradeID::String, units::Union{Real,String}="ALL")

Return an array of trade struct

Arguments

'config::config': a valid struct with user configuracion data

'tradeID::string': a valid trade ID

'units::Union{Number,String}': how much of the Trade to close in units or "ALL"

Examples

closeTrade(userconfig,"66","ALL")

Julianda.Trade.clientExtensions — Type

clientExtensions(config::config, tradeID::String; clientID::String="", tag::String="", comment::String="")

Lets add user information to a specific Trade

Arguments

'config::config': a valid struct with user configuracion data

'tradeID::string': a valid trade ID

clientID, tag and comment: strings with the user information

Example

clientExtensions(userconfig,"66", clientID="007",tag="foo")

Julianda.Trade.setTradeOrders — Function

function setTradeOrders(config::config, tradeID::String; [TP::NamedTuple, SL::NamedTuple, tSL::NamedTuple ])

Create or modify the linked orders for a specific trade

Arguments

'config::config': a valid struct with user configuracion data

'tradeID::string': a valid trade ID

Additional Arguments

'TP::NamedTuple' Take Profit parameters

'SL::NamedTuple': Stop Loss parameters

'tSL::NamedTuple': Trailing Stop Loss parameters

At least one type of order parameters must be provided

Valid order parameters

'price = :Real' :price to create o modify for the specific order. Valir for Stop Loss and Take Profit

'distance = :Real' :price distance to create o modify for the specific order. Valir for Stop Loss and Trailing Stip Loss

'TIF= String': time in force for the order. Valid options are: GTC, GTD, GFD, FOK, IOC. Defaults to GTC

'gtdTime = DateTime': time for GTD (Good unTill Date)

Price and distance are incompatible. Only one can be set for a given order.

Example

setTradeOrders(userconfig, "34"; TP=(price=109.5,), SL=(distance=10,TIF="FOK")) # Do not forget the comma for 1 element NamedTuples

Julianda.Trade.cancelTradeOrders — Function

function cancelTradeOrders(config::config, tradeID::String, orders2cancel::Vector{String})

Cancel linked orders of a specific trade

Arguments

'config::config': a valid struct with user configuracion data

'tradeID::string': a valid trade ID

'orders2cancel::Vector{String}': list of orders to cancel.

order2cancel valid fields are "TP" for Take Profit, "SL" for Stop Loss and "tSL" for Trailing Stop Loss

Example

cancelTradeOrders(userconfig, "34", ["SL", "TP"])

end

Structs

Julianda.Trade.trade — Type

Detailed Trade struct from Oanda

Missing docstring.

Missing docstring for Trade.closeUnitsResp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.clientExtensionsResp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.extensions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.tradeOrdersResponse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.takeProfit. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.stopLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.trailingStopLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.tradeOrders. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Trade.nullTradeOrders. Check Documenter's build log for details.