Transactions

Functions

OandaTrade.Transaction.getTransactionPages — Function

getTransactionPages(config::config; from::DateTime=nothing, to::DateTime=Dates.now(), pageSize::Int=100, type::String=nothing)

Return a struct. Field ':pages' includes the urls for requesting the transactions data in the given timeframe

Arguments

'config::config': a valid config file

'from::Union{DateTime, Nothing}': get transactions from

'to::DateTime': get transactions to

'pageSize::Int': Size of each page

'type::Union{String,Nothing}': type of pages to request

Example

getTransactionPages(userdata, from=DateTime(2019,5,31),type="MARKET_ORDER,STOP_LOSS_ORDER")

OandaTrade.Transaction.getTransaction — Function

getTransaction(config::config, tID::Int)

Return a Dictionary with the transaction data

Arguments

'config::config': a valid config object

'tID::Int': transaction ID

Example

getTransaction(userdata,4)

OandaTrade.Transaction.getTransactions — Function

getTransactions(config::config, fromID::Int, toID::Int, type::Union{String,Nothing}=nothing)
getTransactions(config::config, sinceID::Int, type::Union{String,Nothing}=nothing)

Return an array of Dictionaries with the transactions data

Arguments

'config::config': a valid config struct

'type::Union{String,Nothing}': Specify transaction types

'fromID::Int', 'toID::Int': Get all transactions between two ID's

OR

'sinceID::Int': Get all transactions since an ID

Examples

getTransactions(userdata,3,"MARKET_ORDER,STOP_LOSS_ORDER")
getTransactions(userdata,2,10)

Structs

OandaTrade.Transaction.transactionPages — Type

The transaction pages

Fields

'from': where pages start

'to': where pages end

'pageSize': size of each page

'type::Vector{String}': The type of the transaction pages

'count': number of pages

'pages::Vector{String}': The pages

'lastTransactionID': the last transaction id