Accounts

Functions

Julianda.Account.listAccounts — Function

listAccounts(config::config)

Returns a list of all account IDs and tags authorized for the given Token

Julianda.Account.getAccount — Function

getAccount(config::config)

Returns an Oanda account struct when given a valid config

Julianda.Account.getAccountSummary — Function

getAccountSummary(config::config)

Similar to getAccount but doesnt return the order & trade & positions lists, however it still returns a full account struct, just with these fields left undefined

Julianda.Account.getAccountInstruments — Function

getAccountInstruments(config::config, inst=nothing)

Returns a list of tradeable instruments details for the account

Arguments

inst: Can be left blank to return all tradeable instruments, or as a string csv of instruments to return their details

Julianda.Account.setAccountConfig — Function

setAccountConfig(config::config, alias::String, marginRate::String)

Set client configurable configuration settings

Arguments

alias: The account alias

marginRate: The desired decimal margin rate formatted as a string

Structs

Julianda.Account.account — Type

The account struct given by Oanda

Fields

NAV: The Net Asset Value of an account

alias: User defined alias if one exists

balance: Current Account Balance

createdByUserID: The User ID of the account creator

createdTime: The time the account was created

currency: The primary currency of the account

hedgingEnabled: If the account is allowed to hedge

id: The account ID

lastTransactionID: The last transaction ID

marginAvailable: The margin still available on the account

marginCloseoutMarginUsed: The closeout margin used

marginCloseoutNAV: Margins closeout NAV

marginCloseoutPercent: Margin closeout percent

marginCloseoutPositionValue: Margin closeout position value

marginCloseoutUnrealizedPL: Margin closeout unrealised profit/loss

marginRate: The margin rate

marginUsed: Amount of margin used

openPositionCount: Number of open positions

openTradeCount: Number of open trades

orders: Orders of the account

pendingOrderCount: Number of pending orders

pl: The profit or loss over the lifetime of the account

positionValue: Value of an accounts open positions

positions: Positions of the account

resettablePL: The resetable profit/loss since last reset

trades: Trades of the account

unrealizedPL: The unrealised profit/loss of the account

withdrawalLimit: The withdrawal limit of the account

Julianda.Account.instrumentDetail — Type

Tradeable Instrument data

Fields

displayName: Instrument name

displayPrecision: Decimal precision of the instrument

marginRate: Margin rate on the instrument

maximumOrderUnits: Max units that can be ordered

maximumPositionSize: max position size of the instrument

maximumTrailingStopDistance: max trailing stop distance

minimumTrailingStopDistance: min trailing stop distance

name: Request usable instrument name

pipLocation: current pip location

tradeUnitsPrecision: Decimal precision of trade units

type: Type of instrument