Pricing

Functions

Julianda.Pricing.getPrice — Function

function getPrice(config, instruments)

Get the most recent price update of an instrument

Arguments

- 'config::config': a valid struct with user configuracion data
- 'instruments::Vector{String}': a vector of valid pairs (e.g. ["EUR_USD","EUR_JPY"])

#Example

getPrice(userconfig, ["EUR_USD","EUR_JPY"])

Julianda.Pricing.checkMarket — Function

Get the most recent price update of an instrument

Julianda.Pricing.streamPrice — Function

function streamprice(f, config , instruments)

Returns a stream of price objects and apply a function to each one of them

Arguments

- 'f::Function': a function to apply to each price struct object. Streamprice accepts do block format
- 'config::config': a valid struct with user configuracion data
- 'instruments::Vector{String}': a vector of valid pairs (e.g. ["EUR_USD","EUR_JPY"])

#Example

streamprice(userconfig, ["EUR_JPY"]) do price
    println(price)
end

Missing docstring.

Missing docstring for Pricing.getLatestCandles. Check Documenter's build log for details.

Structs

Julianda.Pricing.priceBucket — Type

Ask / Bid pricing data

Julianda.Pricing.price — Type

Pricing data of an instrument

Missing docstring.

Missing docstring for Pricing.candlestickdata. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Pricing.candlestick. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Pricing.candles. Check Documenter's build log for details.