Transactions

Functions

Julianda.Transaction.getTransactionPages — Function

getTransactionPages(config::config; from::DateTime=nothing, to::DateTime=Dates.now(), pageSize::Int=100, type::String=nothing)

Return a struct. Field ':pages' includes the urls for requesting the transactions data in the given timeframe

Example

getTransactionPages(userdata, from=DateTime(2019,5,31),type="MARKET_ORDER,STOP_LOSS_ORDER")

Julianda.Transaction.getTransaction — Function

getTransaction(config::config, tID::Int)

Return a Dictionary with the transaction data

Example

getTransaction(userdata,4)

Julianda.Transaction.getTransactions — Function

getTransactions(config::config, fromID::Int, toID::Int, type::Union{String,Nothing}=nothing)
getTransactions(config::config, sinceID::Int, type::Union{String,Nothing}=nothing)

Return an array of Dictionaries with the transactions data

Examples

getTransactions(userdata,3,"MARKET_ORDER,STOP_LOSS_ORDER")
getTransactions(userdata,2,10)

Structs

Missing docstring.

Missing docstring for Transaction.transactionPages. Check Documenter's build log for details.